+++ Inc/rtedbg.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Branko Premzel.
+ * Copyright (c) Branko Premzel.
  *
  * SPDX-License-Identifier: MIT
  */
@@ -10,6 +10,17 @@
  * @brief   RTEdbg library function declarations and macro definitions.
  * @note    This file must be included in all source files that use data logging.
  * @version RTEdbg library <DEVELOPMENT BRANCH>
+ *
+ * @note  Add `#define RTE_USE_INLINE_FUNCTIONS` before include `rtedbg.c` if you
+ *        want to use faster inline versions of the logging functions `__rte_msg0()`
+ *        to `__rte_msg4()` in that file.
+ *        If the source file that includes this header file contains interrupt
+ *        programs that have such a high priority that no other interrupt program
+ *        can interrupt them, a faster non-reentrant version of circular buffer
+ *        reservation can be used for logging, allowing even faster (less intrusive)
+ *        logging. This is done by adding the macro
+ *            #define RTE_USE_LOCAL_CPU_DRIVER "rtedbg_generic_non_reentrant.h"
+ *        before #include `rtedbg.c`.
  *******************************************************************************/
 
 #ifndef RTEDBG_H
@@ -21,6 +32,11 @@
 #endif
 #include "rtedbg_config.h"  // Project-specific configuration file.
 #include "rte_system_fmt.h" // System message filter and format ID definitions
+
+#if defined RTE_DISABLE_RTEDBG_FOR_THIS_FILE
+#undef RTE_ENABLED
+#define RTE_ENABLED  0
+#endif
 
 
 #ifdef __cplusplus